테스트코드 분산 및 /users POST 테스트 작성 - main_test.go의 기존 테스트들을 각 endpoint에 맞게 폴더로 분산 - /users POST 작성 및 테스트코드 작성
