+++ src/main_test.go
@@ -1,26 +1,13 @@
 package main
 
 import (
-	"bytes"
-	"encoding/json"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"testing"
 
 	. "github.com/franela/goblin"
-	"github.com/gin-gonic/gin"
+
+	"github.com/json9512/mediumclone-backendwithgo/src/tests"
 )
-
-type data map[string]interface{}
-
-func MakeRequest(r http.Handler, method, path string, body []byte) *httptest.ResponseRecorder {
-	req, _ := http.NewRequest(method, path, bytes.NewBuffer(body))
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	return w
-}
 
 func Test(t *testing.T) {
 	// Setup router
@@ -34,30 +21,33 @@ func Test(t *testing.T) {
 	g.Describe("/posts endpoint tests", func() {
 		// Passing test
 		// GET /ping
-		GETPing(g, router)
+		tests.GETPing(g, router)
 
 		// GET /posts
-		GETPost(g, router)
+		tests.GETPosts(g, router)
 
 		// GET /posts/:id
-		GETPostWithID(g, router)
+		tests.GETPostWithID(g, router)
 
 		// GET /posts/:id/like
-		GETLikesOfPost(g, router)
+		tests.GETLikesOfPost(g, router)
 
 		// GET /posts?tag=rabbit
-		GETPostWithQuery(g, router)
+		tests.GETPostWithQuery(g, router)
 
 		// POST /posts with json {post-id: 5}
-		POSTpostsWithID(g, router)
+		tests.POSTPostWithID(g, router)
 	})
 
 	g.Describe("/users endpoint test", func() {
 		// GET /users
-		GETUsers(g, router)
+		tests.GETUsers(g, router)
 
 		// GET /users/:id
-		GETUsersWithID(g, router)
+		tests.GETUsersWithID(g, router)
+
+		// POST /users
+		tests.POSTUserWithID(g, router)
 	})
 
 	// Environment setup test
@@ -68,178 +58,4 @@ func Test(t *testing.T) {
 		})
 	})
 
-}
-
-func GETPing(g *G, router *gin.Engine) {
-	g.It("GET /ping should return JSON {message: pong}", func() {
-		// Build expected body
-		body := data{
-			"message": "pong",
-		}
-
-		w := MakeRequest(router, "GET", "/ping", nil)
-
-		g.Assert(w.Code).Equal(http.StatusOK)
-
-		var response map[string]string
-
-		err := json.Unmarshal([]byte(w.Body.String()), &response)
-
-		value, exists := response["message"]
-
-		g.Assert(err).IsNil()
-		g.Assert(exists).IsTrue()
-		g.Assert(body["message"]).Equal(value)
-	})
-}
-
-func GETPost(g *G, router *gin.Engine) {
-	g.It("GET /posts should return list of all posts", func() {
-		body := data{
-			"result": []string{"test", "sample", "post"},
-		}
-
-		w := MakeRequest(router, "GET", "/posts", nil)
-
-		g.Assert(w.Code).Eql(http.StatusOK)
-
-		var response map[string][]string
-		err := json.Unmarshal([]byte(w.Body.String()), &response)
-
-		value, exists := response["result"]
-
-		g.Assert(err).IsNil()
-		g.Assert(exists).IsTrue()
-		g.Assert(body["result"]).Eql(value)
-	})
-}
-
-func GETPostWithID(g *G, router *gin.Engine) {
-	g.It("GET /posts/:id should return post with given id", func() {
-		body := data{
-			"result": "5",
-		}
-
-		w := MakeRequest(router, "GET", "/posts/5", nil)
-
-		g.Assert(w.Code).Eql(http.StatusOK)
-
-		var response map[string]string
-		err := json.Unmarshal([]byte(w.Body.String()), &response)
-
-		value, exists := response["result"]
-
-		g.Assert(err).IsNil()
-		g.Assert(exists).IsTrue()
-		g.Assert(body["result"]).Eql(value)
-	})
-}
-
-func GETLikesOfPost(g *G, router *gin.Engine) {
-	g.It("GET /posts/:id/like should return like count of post with given id", func() {
-		body := data{
-			"result": 10,
-		}
-
-		w := MakeRequest(router, "GET", "/posts/5/like", nil)
-
-		g.Assert(w.Code).Eql(http.StatusOK)
-
-		var response map[string]int
-		err := json.Unmarshal([]byte(w.Body.String()), &response)
-
-		value, exists := response["result"]
-
-		g.Assert(err).IsNil()
-		g.Assert(exists).IsTrue()
-		g.Assert(body["result"]).Eql(value)
-	})
-}
-
-func GETPostWithQuery(g *G, router *gin.Engine) {
-	g.It("GET /posts?tag=rabbit should return tags: [rabbit]", func() {
-		tag := make(map[string][]string)
-		tag["tags"] = []string{"rabbit"}
-
-		// build expected body
-		body := data{
-			"result": tag,
-		}
-
-		w := MakeRequest(router, "GET", "/posts?tags=rabbit", nil)
-
-		g.Assert(w.Code).Eql(http.StatusOK)
-
-		var response map[string]map[string][]string
-		err := json.Unmarshal([]byte(w.Body.String()), &response)
-
-		// grab the values
-		value, exists := response["result"]
-
-		g.Assert(err).IsNil()
-		g.Assert(exists).IsTrue()
-		g.Assert(body["result"]).Eql(value)
-	})
-}
-
-func GETUsers(g *G, router *gin.Engine) {
-	g.It("GET /users should return list of all users", func() {
-		body := data{
-			"result": []string{"test", "sample", "users"},
-		}
-
-		w := MakeRequest(router, "GET", "/users", nil)
-
-		g.Assert(w.Code).Eql(http.StatusOK)
-
-		var response map[string][]string
-		err := json.Unmarshal([]byte(w.Body.String()), &response)
-
-		value, exists := response["result"]
-
-		g.Assert(err).IsNil()
-		g.Assert(exists).IsTrue()
-		g.Assert(body["result"]).Eql(value)
-	})
-}
-
-func GETUsersWithID(g *G, router *gin.Engine) {
-	g.It("GET /users/:id should return user with given id", func() {
-		body := data{
-			"result": "5",
-		}
-
-		w := MakeRequest(router, "GET", "/users/5", nil)
-
-		g.Assert(w.Code).Eql(http.StatusOK)
-
-		var response map[string]string
-		err := json.Unmarshal([]byte(w.Body.String()), &response)
-
-		value, exists := response["result"]
-
-		g.Assert(err).IsNil()
-		g.Assert(exists).IsTrue()
-		g.Assert(body["result"]).Eql(value)
-	})
-}
-
-func POSTpostsWithID(g *G, router *gin.Engine) {
-	g.It("POST /posts should create a new post in database", func() {
-		values := data{"post-id": "5"}
-		jsonValue, _ := json.Marshal(values)
-
-		w := MakeRequest(router, "POST", "/posts", jsonValue)
-
-		g.Assert(w.Code).Eql(http.StatusOK)
-
-		var response map[string]string
-		err := json.Unmarshal([]byte(w.Body.String()), &response)
-
-		value, exists := response["post-id"]
-
-		g.Assert(err).IsNil()
-		g.Assert(exists).IsTrue()
-		g.Assert(values["post-id"]).Eql(value)
-	})
 }
